Euler #18
Decided to use a smart algorithm that calculated the maximum path to each node
starting from the top to the bottom. A nodes maximum path is it's value + the
maximum path of it's parents. This is an O(n) algorithm where n is the number of
nodes in the triagle/graph.