Euler #31
An important point is that you don't have to actually produce the different
options but just count them. Also by processing the options in a way that
guarantees that there will be no duplicates sames memory and CPU. I left the
recursion as stupid as possible and then have a bunch of base cases at the
beginning to check for when to stop recursing. That design makes the code really
simple.